fix: upgrade to boto3 for checks
